공통으로 사용 될 함수들 분리
로그인 페이지, 회원가입 페이지에 공통으로 사용할 함수 분리
