+++ signin.html
@@ -40,6 +40,6 @@
         </div>
       </div>
     </main>
-    <script type="module" src="./Scripts/main.js"></script>
+    <script type="module" src="./Scripts/signin.js"></script>
   </body>
 </html>
